msm: vidc: Restrict secure buffer size optimization for higher resolution
For secure usecase, video driver reduces the estimated size by half. This
is done as secure decode spec is in general half of the non secure spec. As
a result, the worst case size get half for secure mode. On Rennel, Based on
this new restriction, the impact on buffer size would be for FWVGA and
below = 6MB, which was earlier 3MB.

CRs-Fixed: 2745960
Change-Id: I4a3be87e3f4beef7e8949cfb85e3341187e4b8b7
Signed-off-by: Vikash Garodia <vgarodia@codeaurora.org>
Signed-off-by: Sanjay Singh <sisanj@codeaurora.org>
Signed-off-by: mARk <r3066.funtab@gmail.com>